requirements.txt
huggingface_hub
datasets
transformers
torch
evaluate
ipywidgets
soundfile
librosa
openai-whisper
jiwer
accelerate
tensorboard
protobuf